requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
allure-python-commons-test==2.9.45
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
distlib==0.3.4
execnet==1.9.0
filelock==3.6.0
flaky==3.7.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
outcome==1.1.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyHamcrest==2.0.3
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
pytest==7.1.1
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==2.0.4
pytest-xdist==2.5.0
requests==2.27.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
virtualenv==20.14.1
webdriver-manager==3.5.4
wsproto==1.1.0

lxml~=4.9.2
cssselect~=1.1.0
yagmail~=0.15.277
cssutils~=2.4.0
cachetools~=5.0.0
premailer~=3.10.0
setuptools~=58.1.0
softest~=1.2.0.0
openpyxl~=3.0.10
Appium-Python-Client~=2.7.1
PyYAML==5.3.1

